Created Testcase for PHPUnit fixtures, which will be used to manage and manipulate fixtures within PHPUnit.
Have finished the first section of test units, which was to focus on creating a method that can retrieve a single and a collection of test data and return it in an array format. There is also functionality allowing us to determine how many pieces of test data are already present within a fixture.

git-svn-id: http://svn.ibetx.com/svn/intranet@49 1f56db89-a82d-46e4-8122-373cd12a268a