Implemented functionality to generate, parse and determine each properties data type.
Refactored the build and drop method to use _runFixtureMethodCall, which now deals with most of the work build & drop previously shared.
 
Also did a little organising, improving on documentation & placing methods in a better order (private, protected, public). This order may need to be discussed as big files like this can be cumbersome to search through.

git-svn-id: http://svn.ibetx.com/svn/intranet@92 1f56db89-a82d-46e4-8122-373cd12a268a