Though FixtureTest was passing all tests, our previous refactoring had broken our PHPUnit_Fixture related tests, mainly due to the fact that whilst refactoring PHPUnitFixturesUnitSuite was not run each time.
Have resolved the above issues as well as carried on the code coverage of PHPUnit_Fixture. Just need to work on privatising PHPUnit_Fixture's properties & possibly accessor for the properties.

Have restructured PHPUnit_Fixture, so the private/protected & public (in that order) methods are together. FixtureTest has been modified to reflect the privatisation of _generateFixtureTestData. As it can not be directly tested we have replaced its tests that cover the same range of code using autoGenerateTestData.

Removed testAutoIncrement, which was being used to determine that if a test data's id was null that it would be autoincremented within the DB, as we know this works now there is no need to keep it, it may be an idea to create functionality that will auto populate these id's
if we are only simulating a query response.

git-svn-id: http://svn.ibetx.com/svn/intranet@115 1f56db89-a82d-46e4-8122-373cd12a268a